brought back the survey page
