+++ app/middleware/02.pageState.global.ts
@@ -1,0 +1,6 @@
+export default defineNuxtRouteMiddleware(async (_to, _from) => {
+  const { setState } = useAppState();
+  const { data: newState } = await useFetch('/api/admin/state');
+
+  setState(newState.value.state);
+});
