+++ app/middleware/requireAuth.ts
@@ -1,0 +1,5 @@
+export default defineNuxtRouteMiddleware(async (_to, _from) => {
+  const { currentUser } = useAuth();
+
+  if (currentUser == null) return abortNavigation();
+});
